Create tg-notify.yml
What the following workflow does is that it sends the user a notification via Telegram when the repo have a new push commit or pull request. The message would contain the commit status (either a push or pull request) with the repo name, followed by the username of the person who triggered the workflow. (Along with the commit SHA & the branch/tag reference that triggered the workflow run)